Deletion of book is only possible if there are no borrow records associated with it with returnstatus 'false'

+++ src/main/java/com/nischal/book_borrowing_app/service/BookService.java
@@ -102,6 +102,11 @@ public class BookService {
 
     @Transactional
     public void deleteBook(Integer id) {
+
+        List<Borrow> borrows = borrowRepository.findByBookId(id);
+        if (!borrows.isEmpty()) {
+            throw new CustomException("Conflict: Cannot delete book it is associated with borrow records.");
+        }
         bookRepository.deleteById(id);
     }
